pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = "tests for omnistrate"
authors = ["FalkorDB <info@falkordb.com>"]
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11.4"
falkordb = {git = "https://github.com/FalkorDB/falkordb-py", rev = "main"}
redis = {git = "https://github.com/dudizimber/redis-py", rev = "master"}
requests = "^2.32.3"
setuptools = "^80.9.0"
curlify = "^3.0.0"
pytest = "^8.3.2"
pytest-timeout = "^2.1.0"
pytest-xdist = "^3.3.0"
pyyaml = "^6.0"
kubernetes = "^27.2.0"
jinja2 = "^3.1.0"
urllib3 = ">=1.24.2,<2.0"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
